CI: Use UUIDs for tests, move to test/unit
Change-type: patch
