--- lib/variables/logger/index.spec.ts
@@ -1,22 +1,0 @@
-import { getEnvironment } from '../../../lib';
-import * as defaults from './defaults';
-
-const variables = {
-	LOGLEVEL: 'crit',
-};
-
-describe('Logger', () => {
-	test('variables are set', () => {
-		const environment = getEnvironment(variables);
-		expect(environment.logger).toEqual({
-			loglevel: variables.LOGLEVEL,
-		});
-	});
-
-	test('defaults are set', () => {
-		const environment = getEnvironment();
-		expect(environment.logger).toEqual({
-			loglevel: defaults.LOGLEVEL,
-		});
-	});
-});
