--- lib/variables/oauth/index.spec.ts
@@ -1,22 +1,0 @@
-import { getEnvironment } from '../../../lib';
-import * as defaults from './defaults';
-
-const variables = {
-	OAUTH_REDIRECT_BASE_URL: 'http://localhost',
-};
-
-describe('OAuth', () => {
-	test('variables are set', () => {
-		const environment = getEnvironment(variables);
-		expect(environment.oauth).toEqual({
-			redirectBaseUrl: variables.OAUTH_REDIRECT_BASE_URL,
-		});
-	});
-
-	test('defaults are set', () => {
-		const environment = getEnvironment();
-		expect(environment.oauth).toEqual({
-			redirectBaseUrl: defaults.OAUTH_REDIRECT_BASE_URL,
-		});
-	});
-});
