+++ package.json
@@ -33,10 +33,10 @@
   "scripts": {
     "clean": "rimraf build",
     "build": "npm run clean && tsc -p tsconfig.build.json",
-    "lint": "balena-lint lib && deplint",
-    "lint:fix": "balena-lint --fix lib",
+    "lint": "balena-lint lib test && deplint",
+    "lint:fix": "balena-lint --fix lib test",
     "test": "npm run lint && npm run test:unit",
-    "test:unit": "jest",
+    "test:unit": "jest test/unit",
     "doc": "typedoc lib/ && touch docs/.nojekyll",
     "prepack": "npm run build"
   },
